+++ test/controllers/dashboard_controller_test.rb
@@ -14,4 +14,24 @@ class DashboardControllerTest < ActionDispatch::IntegrationTest
 
     assert_response :success
   end
+
+  test 'dashboard shows PrusaLink health dot for configured printers' do
+    printer = printers(:prusa_xl)
+    printer.update!(prusalink_key: 'secret', prusalink_reachable: true)
+
+    get root_path
+
+    assert_response :success
+    assert_select '.status-dot.status-success[title=?]', 'PrusaLink connected'
+  end
+
+  test 'dashboard shows red PrusaLink dot when printer is unreachable' do
+    printer = printers(:prusa_xl)
+    printer.update!(prusalink_key: 'secret', prusalink_reachable: false)
+
+    get root_path
+
+    assert_response :success
+    assert_select '.status-dot.status-danger[title=?]', 'PrusaLink unreachable'
+  end
 end